requirements.txt
nltk
textblob
langdetect
pandas
faiss-cpu
